Split tool dispatch into explicit pre-process, execute, and collect phases
Refactors the monolithic tool_dispatch.rs (~1480 lines) into a submodule
directory with focused files for each dispatch phase, improving
maintainability and making each phase independently testable.

Closes #411
